refactor(web): extract useDocumentVisible primitive
