+++ apps/web/src/hooks/usePlatformPresence.ts
@@ -1,17 +1,11 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import { presenceContextForPlatform } from '@kilocode/event-service';
+
+import { useDocumentVisible } from './useDocumentVisible';
 import { usePresenceSubscription } from './usePresenceSubscription';
 
 export function usePlatformPresence() {
-  const [visible, setVisible] = useState(typeof document === 'undefined' ? true : !document.hidden);
-
-  useEffect(() => {
-    const onChange = () => setVisible(!document.hidden);
-    document.addEventListener('visibilitychange', onChange);
-    return () => document.removeEventListener('visibilitychange', onChange);
-  }, []);
-
+  const visible = useDocumentVisible();
   usePresenceSubscription(presenceContextForPlatform('web'), visible);
 }
